fix: fixed ZNA-5595 https://asiatvusa.atlassian.net/browse/ZNA-5595 * updated version
