+++ ZeeHomeScreenManifest.json
@@ -12,7 +12,7 @@
   "platform": "ios",
   "author_name": "O. Liashko",
   "author_email": "o.laksano@applicaster.com",
-  "manifest_version": "2.7.2",
+  "manifest_version": "2.7.3",
   "name": "Zee5HomeScreen",
   "description": "Zee5 Home Screen",
   "type": "general",
@@ -20,7 +20,7 @@
   "identifier": "zee5_home_screen",
   "ui_builder_support": true,
   "dependency_name": "ZeeHomeScreen",
-  "dependency_version": "2.7.2",
+  "dependency_version": "2.7.3",
   "whitelisted_account_ids": [
     "5cb825e102507200087bbdcc",
     "5c3b12762756be0008779106"
